Add files via upload
The jeopardy project requires one to have a solid understanding of dataframe and there are serval stages of processing the data. Data cleaning requires checking column names and renaming column names in a proper format without unexpected spaces, and the basic analysis of the data involves aggregating the data, and the further analyses involve the application of list comprehension, setting up new columns, slicing rows under certain conditions, and defining functions with or without the help of lambda function.